Warning fixes for Clang-3.6
git-svn-id: http://svn.openscenegraph.org/osg/OpenSceneGraph/trunk@15016 16af8721-9629-0410-8352-f15c8da7e697

+++ src/osgPlugins/freetype/FreeTypeLibrary.cpp
@@ -114,7 +114,7 @@ FT_Byte* FreeTypeLibrary::getFace(std::istream& fontstream, unsigned int index, 
     std::streampos length = end - start;
 
     /* empty stream into memory, open that, and keep the pointer in a FreeTypeFont for cleanup */
-    FT_Byte *buffer = new FT_Byte[length];
+    FT_Byte *buffer = new FT_Byte[static_cast<int>(length)];
     fontstream.read(reinterpret_cast<char*>(buffer), length);
     if (!fontstream || (static_cast<std::streampos>(fontstream.gcount()) != length))
     {
